Integração do personagem com o Xp
Agora o personagem pode coletar as bolinhas de Xp e ao serem coletadas elas dão um valor de 15 de Xp. Quando coleta Xp a barra no topo da camera atualiza demonstrando quanto de xp falta para upar.

+++ VampireLike-GGJ/Assets/Scripts/Xp/XpOrb.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class XpOrb : MonoBehaviour
 {
-    [SerializeField] private float value;
+    [SerializeField] private int value;
     private void OnTriggerEnter2D(Collider2D col)
     {
         if (col.CompareTag("Player"))
